services/payment-service/requirements.txt
kafka-python==2.0.2
redis==5.2.1
